Phase 014+015: crosscheck Inputs 전환 + 타입 검증 + OpenAPI path param 충돌 감지 + SSOT 마이그레이션

+++ specs/dummy-gigbridge/func/billing/hold_escrow.go
@@ -5,7 +5,7 @@ package billing
 
 type HoldEscrowRequest struct {
 	GigID    int64
-	Amount   int
+	Amount   int64
 	ClientID int64
 }
